build: don't drop 1 directory level in copy
With the splat, the directory structure is collapsed one level:

    ./this/pkg.spec

becomes:

    ./pkg.spec

This causes issues when an unpacked spec file is present in a
subdirectory as a later glob is used to install the spec's builddeps.
An error is raised by dnf specifically when the additional spec
specifies dependencies that are not installed in the SDK
container. This is the expected behavior and that remains unchanged;
the only thing this change fixes is the inclusion of an unexpected
spec which has unavailable build dependencies.

Signed-off-by: Jacob Vallejo <jakeev@amazon.com>