buildsys: repack github git-archive tarballs
The archive downloads from GitHub use git-archive and are generated at
time of fetch if not recently fetched (depending on their cache). This
means that manifest entries might be added and invalidated if artifacts
are not uploaded to the lookaside cache artifacts "quickly enough".

This change introduces repacked sources specifically for repacking
GitHub sources, impacted by the on-the-fly generation inconsistencies.
The current implementation relies on the host's tooling (GNU variants
specifically) and could easily be performed in native code for improved
portability. This also applies to the use of gzip - as used it *should*
result in the same output; version to version in gzip differences may
challenge this assumption.

Native code may eliminate these issues and is a step easily taken to
implement when desired.

Repack approach taken from the Reproducible Build's website:

https://reproducible-builds.org/docs/archives/

Prior art exists in Nixpkgs where donwloads from sources, like GitHub,
are unpacked to prefer content hashes:

https://github.com/NixOS/nixpkgs/blob/b6f7f1f673046ad2851728cb8f3a2f0888a1dcf0/pkgs/build-support/fetchzip/default.nix#L20

Signed-off-by: Jacob Vallejo <jakeev@amazon.com>